requirements.txt
pydantic>=2.8.0,<3
python-dotenv>=1.0.1,<2
requests>=2.32.0,<3
pandas>=2.2.0,<3
numpy>=1.26,<3


